+++ Software/development/data/com.trivadis.loganalysis.jrockit/src/com/trivadis/loganalysis/jrockit/domain/gc/GarbageCollection.java
@@ -11,13 +11,15 @@
  */
 package com.trivadis.loganalysis.jrockit.domain.gc;
 
+import java.math.BigDecimal;
+
 import com.trivadis.loganalysis.jrockit.domain.State;
 
 public interface GarbageCollection extends Transition {
 
 	String getName();
 
-	double getDuration();
+	BigDecimal getDuration();
 
 	State getStartState();
